Component location refactoring
