+++ src/application/ThemeVars.tsx
@@ -1,31 +1,36 @@
 /** @jsx jsx */
+
 import { useMemo } from "react";
 import { jsx, css, Global } from "@emotion/core";
-import { ColorTheme, Mode, themes } from './theme';
+
+import { ColorTheme, Mode, themes } from "./theme";
 
 const stringifyTheme = (colorTheme: ColorTheme) => {
   const theme = themes[colorTheme];
-  return Object.keys(theme).map(key => `--${key}: ${theme[key]};`).join('');
-}
+  return Object.keys(theme)
+    .map((key) => `--${key}: ${theme[key]};`)
+    .join("");
+};
 
 export const Theme = () => {
   const darkTheme = useMemo(() => stringifyTheme(ColorTheme.Dark), []);
   const lightTheme = useMemo(() => stringifyTheme(ColorTheme.Light), []);
 
   return (
-  <Global
-    styles={css`
-      @media ${Mode.Dark} {
-        :root {
-          ${darkTheme}
+    <Global
+      styles={css`
+        @media ${Mode.Dark} {
+          :root {
+            ${darkTheme}
+          }
         }
-      }
 
-      @media ${Mode.Light} {
-        :root {
-          ${lightTheme}
+        @media ${Mode.Light} {
+          :root {
+            ${lightTheme}
+          }
         }
-      }
-  `} 
-  />);
+      `}
+    />
+  );
 };
