+++ src/application/__tests__/App.test.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import { waitFor } from "@testing-library/react";
+
 import { renderWithApolloClient } from "../utilities/testing/renderWithApolloClient";
-import { currentScreen, Screens } from "../Layouts/Navigation";
+import { currentScreen, Screens } from "../components/Layouts/Navigation";
 import { App, reloadStatus } from "../App";
 
 jest.mock("../Queries/Queries", () => ({
