+++ src/application/components/Cache/common/utils.ts
