+++ src/application/components/Explorer/__mocks__/@forked/graphiql.js
