+++ src/application/index.tsx
@@ -1,33 +1,38 @@
 /** @jsx jsx */
+
 import { useEffect } from "react";
 import { jsx } from "@emotion/core";
 import { render } from "react-dom";
-import { ApolloClient, ApolloProvider, InMemoryCache, useReactiveVar, makeVar, gql } from "@apollo/client";
-import { getOperationName } from "@apollo/client/utilities";
-import { Theme } from "./ThemeVars";
-import "@apollo/space-kit/reset.css";
-
 import {
-  colorTheme,
-  listenForThemeChange,
- } from './theme';
-import { App, reloadStatus } from './App';
-import { resetGraphiQLVars } from './Explorer/Explorer';
+  ApolloClient,
+  ApolloProvider,
+  InMemoryCache,
+  makeVar,
+  gql,
+} from "@apollo/client";
+import { getOperationName } from "@apollo/client/utilities";
+
+import { Theme } from "./ThemeVars";
+import { colorTheme, listenForThemeChange } from "./theme";
+import { App, reloadStatus } from "./App";
+import { resetGraphiQLVars } from "./components/Explorer/Explorer";
+
+import "@apollo/space-kit/reset.css";
 
 const cache = new InMemoryCache({
   typePolicies: {
     WatchedQuery: {
       fields: {
         name(_) {
-          return _ ?? 'Unnamed';
-        }
+          return _ ?? "Unnamed";
+        },
       },
     },
     Mutation: {
       fields: {
         name(_) {
-          return _ ?? 'Unnamed';
-        }
+          return _ ?? "Unnamed";
+        },
       },
     },
     Query: {
@@ -40,7 +45,7 @@ const cache = new InMemoryCache({
         },
         watchedQuery(_, { toReference, args, canRead }) {
           const ref = toReference({
-            __typename: 'WatchedQuery',
+            __typename: "WatchedQuery",
             id: args?.id,
           });
 
@@ -48,16 +53,16 @@ const cache = new InMemoryCache({
         },
         mutation(_, { toReference, args }) {
           return toReference({
-            __typename: 'Mutation',
+            __typename: "Mutation",
             id: args?.id,
           });
         },
         cache() {
           return cacheVar();
-        }
-      }
-    }
-  }
+        },
+      },
+    },
+  },
 });
 
 const cacheVar = makeVar(null);
@@ -99,15 +104,15 @@ interface Query {
 }
 
 type WatchedQuery = Query & {
-  __typename: 'WatchedQuery';
+  __typename: "WatchedQuery";
   queryString: string;
   cachedData: object;
-}
+};
 
 type Mutation = Query & {
-  __typename: 'Mutation';
+  __typename: "Mutation";
   mutationString: string;
-}
+};
 
 export function getQueryData(query, key: number): WatchedQuery | undefined {
   if (!query) return;
@@ -115,18 +120,18 @@ export function getQueryData(query, key: number): WatchedQuery | undefined {
   // We need a workaround to show that data + we should surface
   // the FetchPolicy.
   const name = getOperationName(query?.document);
-  if (name === 'IntrospectionQuery') {
+  if (name === "IntrospectionQuery") {
     return;
   }
 
   return {
     id: key,
-    __typename: 'WatchedQuery',
+    __typename: "WatchedQuery",
     name,
     queryString: query?.source?.body,
     variables: query.variables,
     cachedData: query.cachedData,
-  }
+  };
 }
 
 export function getMutationData(mutation, key: number): Mutation | undefined {
@@ -134,15 +139,17 @@ export function getMutationData(mutation, key: number): Mutation | undefined {
 
   return {
     id: key,
-    __typename: 'Mutation',
+    __typename: "Mutation",
     name: getOperationName(mutation?.document),
     mutationString: mutation?.source?.body,
     variables: mutation.variables,
-  }
+  };
 }
 
 export const writeData = ({ queries, mutations, cache }) => {
-  const filteredQueries: WatchedQuery[] = queries.map((q, i: number) => getQueryData(q, i)).filter(Boolean);
+  const filteredQueries: WatchedQuery[] = queries
+    .map((q, i: number) => getQueryData(q, i))
+    .filter(Boolean);
 
   client.writeQuery({
     query: GET_QUERIES,
@@ -150,19 +157,21 @@ export const writeData = ({ queries, mutations, cache }) => {
       watchedQueries: {
         queries: filteredQueries,
         count: filteredQueries.length,
-      }
+      },
     },
   });
 
-  const mappedMutations: Mutation[] = mutations.map((m, i: number) => getMutationData(m, i));
-  
+  const mappedMutations: Mutation[] = mutations.map((m, i: number) =>
+    getMutationData(m, i)
+  );
+
   client.writeQuery({
     query: GET_MUTATIONS,
     data: {
       mutationLog: {
         mutations: mappedMutations,
         count: mappedMutations.length,
-      }
+      },
     },
   });
   cacheVar(cache);
@@ -179,7 +188,9 @@ export const handleReloadComplete = () => {
 };
 
 export const AppProvider = () => {
-  useEffect(() => listenForThemeChange((newColorTheme) => colorTheme(newColorTheme)));
+  useEffect(() =>
+    listenForThemeChange((newColorTheme) => colorTheme(newColorTheme))
+  );
 
   return (
     <ApolloProvider client={client}>
@@ -190,7 +201,5 @@ export const AppProvider = () => {
 };
 
 export const initDevTools = () => {
-  render(<AppProvider />,
-    document.getElementById("devtools")
-  );
+  render(<AppProvider />, document.getElementById("devtools"));
 };
