+++ src/extension/devtools/panel.ts
@@ -1,5 +1,13 @@
-import { initDevTools, writeData, handleReload, handleReloadComplete } from "../../application";
-import { receiveGraphiQLRequests, sendResponseToGraphiQL } from "../../application/Explorer/graphiQLRelay";
+import {
+  initDevTools,
+  writeData,
+  handleReload,
+  handleReloadComplete,
+} from "../../application";
+import {
+  receiveGraphiQLRequests,
+  sendResponseToGraphiQL,
+} from "../../application/components/Explorer/graphiQLRelay";
 
 (window as any).__DEVTOOLS_APPLICATION__ = {
   initialize: initDevTools,
